started migrating default query to default_query.py in the DefaultQuery class
